+++ intellij-plugin-structure/structure-toolbox/src/main/kotlin/com/jetbrains/plugin/structure/toolbox/ToolboxPlugin.kt
@@ -5,6 +5,7 @@
 package com.jetbrains.plugin.structure.toolbox
 
 import com.jetbrains.plugin.structure.base.plugin.Plugin
+import com.jetbrains.plugin.structure.base.plugin.PluginFile
 import com.jetbrains.plugin.structure.base.plugin.PluginIcon
 import com.jetbrains.plugin.structure.base.plugin.ThirdPartyDependency
 
@@ -24,24 +25,4 @@ data class ToolboxPlugin(
   override val vendorEmail: String? = null
   override val vendorUrl: String? = null
   override val url: String? = null
-}
-
-data class PluginFile(val name: String, val content: ByteArray) {
-  override fun equals(other: Any?): Boolean {
-    if (this === other) return true
-    if (javaClass != other?.javaClass) return false
-
-    other as PluginFile
-
-    if (name != other.name) return false
-    if (!content.contentEquals(other.content)) return false
-
-    return true
-  }
-
-  override fun hashCode(): Int {
-    var result = name.hashCode()
-    result = 31 * result + content.contentHashCode()
-    return result
-  }
 }
